refactor: move category creator component
